requirements.txt
Flask>=0.12
twilio~=6.0.0
dialogflow>=0.4.0
requests==2.18.4
google-cloud-vision==0.34.0
PyMySQL==0.9.2